+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.1.1
+
+* The error message for canceling a task is fixed
+* fix error prompts
+
 ## 0.1.0
 
 * initialVersion.